table: split style.go into individual files (#392)
